Item13321: not an ideal solution, but it's what the rcs store does so can't be any worse. The Foswiki::Meta code is actually OK with this kind of weak lock

+++ PlainFileStoreContrib/lib/Foswiki/Store/PlainFile.pm
@@ -596,8 +596,10 @@ sub atomicUnlock {
     my ( $this, $meta, $cUID ) = @_;
 
     my $filename = _getData($meta) . '.lock';
-    unlink($filename)
-      or die "PlainFile: failed to delete $filename: $!";
+    if ( -e $filename ) {
+        unlink($filename)
+          or die "PlainFile: failed to delete $filename: $!";
+    }
 }
 
 # Implement Foswiki::Store
